add Retrofit and Flow
сделано:
- загрузка списка для первого экрана.
- загрузка детальной информации для второго экрана.
Нужно добавить:
- параллельную загрузку данных,
- прогрессБар на время загрузки
- обработку отсутствия интернета

+++ app/src/main/java/ru/myacademyhomework/myfilms/data/Movie.kt
@@ -18,7 +18,7 @@ data class Movie(
     @SerialName("vote_average")
     val ratings: Float,
     @SerialName("vote_count")
-    val numberOfRatings: Int,
+    val numberOfRatings: Long,
 
     val minimumAge: Int,
